Move diagnostic functions into the `miette` module.
This change moves `Glob` functions enabled by the `miette` feature into
the `miette` module (much like functions enabled by the `walk` feature).
