+++ crates/turborepo-wax/src/diagnostics/miette.rs
@@ -3,13 +3,68 @@
 use std::borrow::Cow;
 
 use miette::{Diagnostic, SourceSpan};
-use tardar::BoxedDiagnostic;
+use tardar::{
+    BoxedDiagnostic, DiagnosticResult, DiagnosticResultExt as _, IteratorExt as _, ResultExt as _,
+};
 use thiserror::Error;
 
 use crate::{
     diagnostics::SpanExt as _,
+    rule,
     token::{self, TokenKind, TokenTree, Tokenized},
+    Checked, Glob,
 };
+
+/// APIs for diagnosing globs.
+impl<'t> Glob<'t> {
+    /// Constructs a [`Glob`] from a glob expression with diagnostics.
+    ///
+    /// This function is the same as [`Glob::new`], but additionally returns
+    /// detailed diagnostics on both success and failure.
+    ///
+    /// See [`Glob::diagnose`].
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// use tardar::DiagnosticResultExt as _;
+    /// use wax::Glob;
+    ///
+    /// let result = Glob::diagnosed("(?i)readme.{md,mkd,markdown}");
+    /// for diagnostic in result.diagnostics() {
+    ///     eprintln!("{}", diagnostic);
+    /// }
+    /// if let Some(glob) = result.ok_output() { /* ... */ }
+    /// ```
+    ///
+    /// [`Glob`]: crate::Glob
+    /// [`Glob::diagnose`]: crate::Glob::diagnose
+    /// [`Glob::new`]: crate::Glob::new
+    pub fn diagnosed(expression: &'t str) -> DiagnosticResult<'t, Self> {
+        parse_and_diagnose(expression).and_then_diagnose(|tree| {
+            Glob::compile(tree.as_ref().tokens())
+                .into_error_diagnostic()
+                .map_output(|program| Glob { tree, program })
+        })
+    }
+
+    /// Gets **non-error** [`Diagnostic`]s.
+    ///
+    /// This function requires a receiving [`Glob`] and so does not report
+    /// error-level [`Diagnostic`]s. It can be used to get non-error
+    /// diagnostics after constructing or [partitioning][`Glob::partition`]
+    /// a [`Glob`].
+    ///
+    /// See [`Glob::diagnosed`].
+    ///
+    /// [`Diagnostic`]: miette::Diagnostic
+    /// [`Glob`]: crate::Glob
+    /// [`Glob::diagnosed`]: crate::Glob::diagnosed
+    /// [`Glob::partition`]: crate::Glob::partition
+    pub fn diagnose(&self) -> impl Iterator<Item = Box<dyn Diagnostic + '_>> {
+        diagnose(self.tree.as_ref())
+    }
+}
 
 #[derive(Clone, Debug, Diagnostic, Error)]
 #[diagnostic(code(wax::glob::semantic_literal), severity(warning))]
@@ -32,7 +87,19 @@ pub struct TerminatingSeparatorWarning<'t> {
     span: SourceSpan,
 }
 
-pub fn diagnose<'i, 't>(
+fn parse_and_diagnose(expression: &str) -> DiagnosticResult<Checked<Tokenized>> {
+    token::parse(expression)
+        .into_error_diagnostic()
+        .and_then_diagnose(|tokenized| rule::check(tokenized).into_error_diagnostic())
+        .and_then_diagnose(|checked| {
+            // TODO: This should accept `&Checked`.
+            diagnose(checked.as_ref())
+                .into_non_error_diagnostic()
+                .map_output(|_| checked)
+        })
+}
+
+fn diagnose<'i, 't>(
     tokenized: &'i Tokenized<'t>,
 ) -> impl 'i + Iterator<Item = BoxedDiagnostic<'t>> {
     None.into_iter()
@@ -68,7 +135,6 @@ pub fn diagnose<'i, 't>(
         )
 }
 
-// These tests exercise `Glob` APIs, which wrap functions in this module.
 #[cfg(test)]
 mod tests {
     use crate::Glob;
